+++ src/main/java/ortus/boxlang/runtime/loader/resolvers/BoxResolver.java
@@ -17,9 +17,7 @@
  */
 package ortus.boxlang.runtime.loader.resolvers;
 
-import java.io.File;
 import java.io.IOException;
-import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 import java.util.ArrayList;
@@ -36,6 +34,7 @@ import ortus.boxlang.runtime.loader.ImportDefinition;
 import ortus.boxlang.runtime.runnables.RunnableLoader;
 import ortus.boxlang.runtime.scopes.Key;
 import ortus.boxlang.runtime.types.IStruct;
+import ortus.boxlang.runtime.util.FileSystemUtil;
 import ortus.boxlang.runtime.util.ResolvedFilePath;
 
 /**
@@ -47,11 +46,6 @@ public class BoxResolver extends BaseResolver {
 	 * Singleton instance
 	 */
 	protected static BoxResolver				instance;
-
-	/**
-	 * Flag for whether FS is case sensitive or not to short circuit case insensitive path resolution
-	 */
-	private static boolean						isCaseSensitiveFS	= caseSensitivityCheck();
 
 	/**
 	 * List of valid class extensions
@@ -217,7 +211,7 @@ public class BoxResolver extends BaseResolver {
 				    Path absolutePath = Path.of( StringUtils.replaceOnceIgnoreCase( slashName, entry.getKey().getName(), entry.getValue() + "/" ) + extension )
 				        .normalize();
 				    // Verify that the file exists
-				    absolutePath = pathExists( absolutePath );
+				    absolutePath = FileSystemUtil.pathExistsCaseInsensitive( absolutePath );
 				    if ( absolutePath != null ) {
 					    try {
 						    String mappingName		= entry.getKey().getName();
@@ -322,92 +316,12 @@ public class BoxResolver extends BaseResolver {
 		for ( String extension : VALID_EXTENSIONS ) {
 			Path	targetPath	= parentPath.resolve( slashName.substring( 1 ) + extension ).normalize();
 
-			Path	result		= pathExists( targetPath );
+			Path	result		= FileSystemUtil.pathExistsCaseInsensitive( targetPath );
 			if ( result != null ) {
 				return result;
 			}
 		}
 		return null;
-	}
-
-	/**
-	 * Performs case insensitive path resolution
-	 * 
-	 * @param path The path to check
-	 * 
-	 * @return The resolved path or null if not found
-	 */
-	Path pathExists( Path path ) {
-		Boolean defaultCheck = Files.exists( path );
-		if ( defaultCheck ) {
-			try {
-				return path.toRealPath();
-			} catch ( IOException e ) {
-				return null;
-			}
-		}
-		if ( isCaseSensitiveFS ) {
-			String		realPath		= "";
-			String[]	pathSegments	= path.toString().replace( '\\', '/' ).split( "/" );
-			if ( pathSegments.length > 0 && pathSegments[ 0 ].contains( ":" ) ) {
-				realPath = pathSegments[ 0 ];
-			}
-			Boolean first = true;
-			for ( String thisSegment : pathSegments ) {
-				// Skip windows drive letter
-				if ( realPath == pathSegments[ 0 ] && pathSegments[ 0 ].contains( ":" ) && first ) {
-					first = false;
-					continue;
-				}
-				// Skip empty segments
-				if ( thisSegment.length() == 0 ) {
-					continue;
-				}
-
-				Boolean		found		= false;
-				String[]	children	= new File( realPath + "/" ).list();
-				// This will happen if we have a matched file in the middle of a path like /foo/index.cfm/bar
-				if ( children == null ) {
-					return null;
-				}
-				for ( String thisChild : children ) {
-					// We're taking the FIRST MATCH. Buyer beware
-					if ( thisSegment.equalsIgnoreCase( thisChild ) ) {
-						realPath	+= "/" + thisChild;
-						found		= true;
-						break;
-					}
-				}
-				// If we made it through the inner loop without a match, we've hit a dead end
-				if ( !found ) {
-					return null;
-				}
-			}
-			// If we made it through the outer loop, we've found a match
-			Path realPathFinal = Paths.get( realPath );
-			return realPathFinal;
-		}
-		return null;
-	}
-
-	private static boolean caseSensitivityCheck() {
-		try {
-			File	currentWorkingDir	= new File( System.getProperty( "user.home" ) );
-			File	case1				= new File( currentWorkingDir, "case1" );
-			File	case2				= new File( currentWorkingDir, "Case1" );
-			case1.createNewFile();
-			if ( case2.createNewFile() ) {
-				case1.delete();
-				case2.delete();
-				return true;
-			} else {
-				case1.delete();
-				return false;
-			}
-		} catch ( Throwable e ) {
-			e.printStackTrace();
-		}
-		return true;
 	}
 
 }
