+++ src/main/java/ortus/boxlang/runtime/runnables/RunnableLoader.java
@@ -17,6 +17,7 @@
  */
 package ortus.boxlang.runtime.runnables;
 
+import java.nio.file.Path;
 import java.util.Set;
 
 import ortus.boxlang.compiler.IBoxpiler;
@@ -112,12 +113,39 @@ public class RunnableLoader {
 	 * @return
 	 */
 	public BoxTemplate loadTemplateAbsolute( IBoxContext context, ResolvedFilePath resolvedFilePath ) {
-		// TODO: Make case insensitive
-		if ( !resolvedFilePath.absolutePath().toFile().exists() ) {
+		Path result = FileSystemUtil.pathExistsCaseInsensitive( resolvedFilePath.absolutePath() );
+		if ( result == null ) {
 			throw new MissingIncludeException( "The template path [" + resolvedFilePath.absolutePath().toString() + "] could not be found.",
 			    resolvedFilePath.absolutePath().toString() );
 		}
+		// If the path found on disk is not the same as the resolved path, then we need to update the resolved path
+		// This would happen on a case-sensitive file system where the incoming path had the incorrect case, but we
+		// we want our resolvedPath instance to reflect the real path on disk.
+		if ( !result.equals( resolvedFilePath.absolutePath() ) ) {
+			// relative path can be null
+			int lengthToUse = resolvedFilePath.relativePath() != null
+			    ? resolvedFilePath.relativePath().length()
+			    : resolvedFilePath.absolutePath().toString().length();
 
+			// Check if the result length is less than or equal to the length to use
+			if ( result.toString().length() <= lengthToUse ) {
+				// Create new relative path with last N chars of the new path
+				resolvedFilePath = ResolvedFilePath.of(
+				    resolvedFilePath.mappingName(),
+				    resolvedFilePath.mappingPath(),
+				    resolvedFilePath.absolutePath().toString().substring( result.toString().length() ),
+				    result
+				);
+			} else {
+				// Reuse the old relative path
+				resolvedFilePath = ResolvedFilePath.of(
+				    resolvedFilePath.mappingName(),
+				    resolvedFilePath.mappingPath(),
+				    resolvedFilePath.relativePath(),
+				    result
+				);
+			}
+		}
 		String	ext			= "";
 		String	fileName	= resolvedFilePath.absolutePath().getFileName().toString().toLowerCase();
 		if ( fileName.contains( "." ) ) {
